Time: 8 ms (59.66%), Space: 22.9 MB (76.88%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -6,7 +6,7 @@ public:
 
         while (l <= r) {
             int mid = (l + r) / 2;
-            int hours = 0;
+            double hours = 0;
             for (int p : piles)
                 hours += (p + mid - 1) / mid;
